BAU: Make working dir optional
Make the working dir param optional so the calling workflow doesn't need to provide it if there's no need to cd from the repo root.

Remove the cd step as it has no effect on the following step. The effect of the cd command isn't carried between steps - changing the dir is achieved by specifying the `working-directory` property on the step, which is already set for the `Upload Lambdas to S3` step.